Update ConvertTo-PoshTools.ps1
in addition to the color object and the custom color array, there's the third option of a systemcolors object.  This edit will try to cast the value as systemcolors, then colors, then finally just create a color object from a custom array.  This preserves the class type in visual studio which looks like it might be important for the systemcolors stuff